requirements.txt
anyio>=4.0.0
asgiref>=3.7.2
bcrypt>=4.0.1
cffi>=1.16.0
click>=8.1.7
cryptography>=41.0.7
dnspython>=2.4.2
ecdsa>=0.18.0
email-validator>=2.1.0
fastapi>=0.109.0
h11>=0.14.0
idna>=3.6
motor>=3.3.2
passlib>=1.7.4
pyasn1>=0.5.1
pycparser>=2.21
pydantic>=1.10.13,<2.0.0
PyJWT>=2.8.0
pymongo>=4.6.1
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
PyYAML>=6.0.1
rsa>=4.9
sniffio>=1.3.0
starlette>=0.36.0
typing_extensions>=4.9.0
uvicorn>=0.27.0
